Extract performance confirmation modal into separate method, cleaned up performance popups hook
